clang errors/warnings

+++ trunk/linbox/tests/test-inverse.C
@@ -6,20 +6,20 @@
  *
  * --------------------------------------------------------
  *
- * 
+ *
  * ========LICENCE========
  * This file is part of the library LinBox.
- * 
+ *
  * LinBox is free software: you can redistribute it and/or modify
  * it under the terms of the  GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 2.1 of the License, or (at your option) any later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
@@ -447,7 +447,7 @@ int main (int argc, char **argv)
 
 	static size_t n = 10;
 	static integer q = 2147483647U;
-	static int iterations = 100;
+	static unsigned int iterations = 100;
 	static int N = 1;
 
 	static Argument args[] = {
@@ -471,7 +471,7 @@ int main (int argc, char **argv)
 	commentator().getMessageClass (INTERNAL_DESCRIPTION).setMaxDepth (3);
 
 	RandomDenseStream<Field, Vector> stream1 (F, n, iterations), stream2 (F, n, iterations);
-	RandomDenseStream<Field, Vector> stream3 (F, n, N);
+	RandomDenseStream<Field, Vector> stream3 (F, n, (size_t)N);
 
 	if (!testIdentityInverse    (F, stream1)) pass = false;
 	if (!testVandermondeInverse (F, stream1, stream3)) pass = false;
